+++ ViroRenderer/VROLayerSubstrate.h
@@ -33,7 +33,7 @@ public:
     virtual void render(const VRORenderContext &context,
                         matrix_float4x4 mv,
                         vector_float4 bgColor) = 0;
-    virtual void setContents(const void *data, const size_t dataLength, int width, int height) = 0;
+    virtual void setContents(const void *data, const size_t dataLength, size_t width, size_t height) = 0;
 
 protected:
